+++ src/renderer/components/InputColor/InputColor.tsx
@@ -5,13 +5,15 @@ import Input from '../Input';
 
 import styles from './InputColor.module.scss';
 
+const COLOR_CHARS = '0 1 2 3 4 5 6 7 8 9 a b c d e f'.split(' ');
+
 const cx = classNames.bind(styles);
 
 type InputProps = Omit<React.HTMLProps<HTMLInputElement>, 'onChange' | 'value'>;
 type Props = InputProps & {
   className?: string;
   id: string;
-  value: string | number | null;
+  value: string | null;
   onChange: (value: string) => unknown;
   block?: boolean;
 };
@@ -29,11 +31,34 @@ const InputColor: React.FC<Props> = ({
   block,
   ...rest
 }) => {
-  const handleChange = useCallback(
-    (event) => {
+  const handleKeyPress = useCallback(
+    (event: React.KeyboardEvent<HTMLInputElement>) => {
+      if (!value && event.key === '#') return;
+
+      if (
+        value &&
+        value.startsWith('#') &&
+        value.length < 7 &&
+        COLOR_CHARS.includes(event.key.toLowerCase())
+      )
+        return;
+      event.preventDefault();
+    },
+    [value]
+  );
+
+  const handleColorChange = useCallback(
+    (event: React.ChangeEvent<HTMLInputElement>) => {
       const { value: newValue } = event.target;
 
-      onChange(newValue);
+      onChange(newValue.toLowerCase());
+    },
+    [onChange]
+  );
+
+  const handleTextChange = useCallback(
+    (newValue: string) => {
+      onChange(newValue.toLowerCase());
     },
     [onChange]
   );
@@ -43,12 +68,13 @@ const InputColor: React.FC<Props> = ({
       className={cx('base', className)}
       id={id}
       value={value}
-      onChange={onChange}
+      onKeyPress={handleKeyPress}
+      onChange={handleTextChange}
       left={
         <input
           className={cx('color')}
           type="color"
-          onChange={handleChange}
+          onChange={handleColorChange}
           value={value ?? ''}
         />
       }