+++ src/renderer/hooks/useNotes.ts
@@ -8,6 +8,11 @@ import {
   getMidiNote,
   getMidiValue,
 } from 'renderer/helpers';
+import {
+  getKeySignature,
+  getNoteInKeySignature,
+  KeySignatureConfig,
+} from 'renderer/helpers/note';
 import useMidiMessage from './useMidiMessage';
 
 const MIDI_CMD_NOTE_OFF = 0x80;
@@ -16,12 +21,12 @@ const MIDI_CHANNEL_ALL = 0;
 const MIDI_CMD_CC = 0xb0;
 const MIDI_CC_SUSTAIN = 0x40;
 
-const getChordInfo = (chord: string) => {
+const getChordInfo = (chord: string, keySignatureNotes: string[]) => {
   const match = chord.match(/^([A-G](b|#)?)([^/]+)(\/([A-G](b|#)?))?$/);
   if (match) {
-    const tonic = match[1];
+    const tonic = getNoteInKeySignature(match[1], keySignatureNotes);
     const type = match[3];
-    const root = match[5];
+    const root = getNoteInKeySignature(match[5], keySignatureNotes);
     const c = Chord.getChord(type, tonic);
     const rootInterval = Interval.distance(tonic, root);
     const rootDegree = c.intervals.indexOf(rootInterval) + 1;
@@ -30,14 +35,19 @@ const getChordInfo = (chord: string) => {
   return null;
 };
 
-const getChords = (notes: string[]) => {
-  const chords = Chord.detect(notes).map(getChordInfo);
+const getChords = (notes: string[], keySignatureNotes: string[]) => {
+  const chords = Chord.detect(notes).map((n) =>
+    getChordInfo(n, keySignatureNotes)
+  );
 
   return chords;
 };
 
+const midiSortCompareFn = (a: number, b: number) => a - b;
+
 interface Parameters {
   accidentals: 'flat' | 'sharp';
+  key: string;
   midiChannel: number;
 }
 
@@ -55,7 +65,7 @@ interface MidiAction {
 }
 
 enum ParametersActionTypes {
-  ACCIDENTALS_CHANGED = 'ACCIDENTALS_CHANGED',
+  KEY_SIGNATURE_CHANGED = 'KEY_SIGNATURE_CHANGED',
 }
 interface ParametersAction {
   type: ParametersActionTypes;
@@ -70,40 +80,36 @@ interface State {
   playedMidiNotes: number[];
   midiNotes: number[];
   notes: string[];
+  pitchClasses: string[];
   chords: ReturnType<typeof getChords>;
   sustained: boolean;
-  params: {
-    accidentals: 'flat' | 'sharp';
-  };
+  keySignature: KeySignatureConfig;
 }
 
 // Our reducer function that uses a switch statement to handle our actions
 function reducer(state: State, action: Action): State {
   const { type } = action;
 
-  const { accidentals } = state.params;
+  const { notes: keySignatureNotes } = state.keySignature;
 
-  const fromMidi =
-    accidentals === 'sharp' ? Note.fromMidiSharps : Note.fromMidi;
+  const fromMidi = (m: number) =>
+    getNoteInKeySignature(Note.fromMidi(m), keySignatureNotes);
 
   switch (type) {
-    case ParametersActionTypes.ACCIDENTALS_CHANGED: {
-      const newAccidentals = action.value as typeof state.params['accidentals'];
-      const notes = state.midiNotes.map((m) =>
-        Note.pitchClass(
-          newAccidentals === 'sharp' ? Note.fromMidiSharps(m) : Note.fromMidi(m)
-        )
+    case ParametersActionTypes.KEY_SIGNATURE_CHANGED: {
+      const keySignature = action.value as typeof state.keySignature;
+      const notes = state.midiNotes.map((m: number) =>
+        getNoteInKeySignature(Note.fromMidi(m), keySignature.notes)
       );
-      const chords = getChords(notes);
+      const pitchClasses = notes.map(Note.pitchClass);
+      const chords = getChords(notes, keySignature.notes);
 
       return {
         ...state,
         notes,
+        pitchClasses,
         chords,
-        params: {
-          ...state.params,
-          accidentals: newAccidentals,
-        },
+        keySignature,
       };
     }
     case MidiActionTypes.NOTE_ON: {
@@ -113,13 +119,15 @@ function reducer(state: State, action: Action): State {
         (m) => m !== midi
       );
       const midiNotes = [...sustainedMidiNotes, ...playedMidiNotes];
-      midiNotes.sort();
-      const notes = midiNotes.map((m) => Note.pitchClass(fromMidi(m)));
-      const chords = getChords(notes);
+      midiNotes.sort(midiSortCompareFn);
+      const notes = midiNotes.map(fromMidi);
+      const pitchClasses = notes.map(Note.pitchClass);
+      const chords = getChords(notes, keySignatureNotes);
 
       return {
         ...state,
         notes,
+        pitchClasses,
         midiNotes,
         sustainedMidiNotes,
         playedMidiNotes,
@@ -133,12 +141,15 @@ function reducer(state: State, action: Action): State {
         ? [...state.sustainedMidiNotes, midi]
         : state.sustainedMidiNotes;
       const midiNotes = [...sustainedMidiNotes, ...playedMidiNotes];
-      midiNotes.sort();
-      const notes = midiNotes.map((m) => Note.pitchClass(fromMidi(m)));
-      const chords = getChords(notes);
+      midiNotes.sort(midiSortCompareFn);
+      const notes = midiNotes.map(fromMidi);
+      const pitchClasses = notes.map(Note.pitchClass);
+      const chords = getChords(notes, keySignatureNotes);
+
       return {
         ...state,
         notes,
+        pitchClasses,
         playedMidiNotes,
         midiNotes,
         sustainedMidiNotes,
@@ -156,15 +167,17 @@ function reducer(state: State, action: Action): State {
 
     case MidiActionTypes.SUSTAIN_OFF: {
       const midiNotes = [...state.playedMidiNotes];
-      midiNotes.sort();
-      const notes = midiNotes.map((m) => Note.pitchClass(Note.fromMidi(m)));
-      const chords = getChords(notes);
+      midiNotes.sort(midiSortCompareFn);
+      const notes = midiNotes.map(fromMidi);
+      const pitchClasses = notes.map(Note.pitchClass);
+      const chords = getChords(notes, keySignatureNotes);
       return {
         ...state,
         sustained: false,
         sustainedMidiNotes: [],
         midiNotes,
         notes,
+        pitchClasses,
         chords,
       };
     }
@@ -179,30 +192,28 @@ const defaultState: State = {
   playedMidiNotes: [],
   midiNotes: [],
   notes: [],
+  pitchClasses: [],
   chords: [],
   sustained: false,
-  params: {
-    accidentals: 'flat',
-  },
+  keySignature: getKeySignature('C'),
 };
 
 export default function useNotes({
   accidentals = 'flat',
+  key = 'C',
   midiChannel = MIDI_CHANNEL_ALL,
 }: Partial<Parameters> = {}) {
   const [state, dispatch] = useReducer(reducer, {
     ...defaultState,
-    params: {
-      accidentals,
-    },
+    keySignature: getKeySignature(key, accidentals === 'sharp'),
   });
 
   useEffect(() => {
     dispatch({
-      type: ParametersActionTypes.ACCIDENTALS_CHANGED,
-      value: accidentals,
+      type: ParametersActionTypes.KEY_SIGNATURE_CHANGED,
+      value: getKeySignature(key, accidentals === 'sharp'),
     });
-  }, [accidentals]);
+  }, [accidentals, key]);
 
   const onMidiMessage = useCallback(
     (message: MidiMessage) => {