+++ CHANGELOG.md
@@ -1,6 +1,28 @@
-# 1.0.0 (2022-05-30)
+# 1.1.0 (2022-08-07)
 
-#### Features
+## Features
 
-- **MIDI:** discover, connect, routing
-- **Chord Display:** detect chord, display piano with notes
+- **Chord Display:** Added music notation with VexFlow
+- **Chord Display:** Added Key Signature with note names in key
+- **Settings:** Input for Notes + Midi Learn
+- **Debugger:** Filter Midi Clock
+- **Overlay:** auto reconnect websocket
+- **chore:** Added Linux AppImage build + fix traffic lights
+
+## Fixes
+
+- **Settings:** Fixed Input Color accepting any text
+- **Chord Display:** Fixed Keyboard size overflowing in some windows sizes
+- **Debugger:** Fixed message manager being disposed in useMidiMessages hook
+- **Home:** Wrong link to report issues
+- **chore:** upgrade dependencies and Electron v19
+
+# 1.0.0 (2022-07-26)
+
+## Features
+
+- **MIDI:** discover, connect, routing, latency monitoring
+- **Chord Display:** Detect chord, display piano with notes with sustain, UI customization
+- **Overlay:** HTTP/WS server for external integration
+- **Debugger:** Display All MIDI messages received
+- **General:** Launch at startup, Always on Top + Window interactions